\name{topo_dem}
\docType{data}
\alias{digital elevation}
\title{matrix of digital elevation}
\description{A small matrix that is representative of a digital elevation model. Row and column positions take the place of spatial coordinates}
\usage{data(topo_dem)}
\format{
\code{topo_dem} is an 109 row \code{matrix} with 110 columns. Values correspond to elevation. 
}
\details{This dataset is used to exemplfy the procedure for generating random catena or toposequences which is described in the Using R for Digital Soil Mapping book.}

\references{
\itemize{
\item Malone, B.P., Minasny, B., McBratney, A.B. (2017) \href{https://link.springer.com/book/10.1007/978-3-319-44327-0}{Using R for Digital Soil Mapping}. Springer Cham. 262 Pages.
}
}

\examples{

# library(ithir)
# data(topo_dem)
# str(topo_dem)

}
\keyword{datasets}
