\name{USYD drainage index data}
\docType{data}
\alias{USYD soil drainage data}
\title{Soil drainage index data}
\description{A \code{dataframe} of 446 rows. Each row corresponds to a location. Actual location not provided. The \code{dataframe} has 2 variables: \code{DI_observed} and \code{DI_predicted}.}
\usage{data(USYD_dIndex)}
\format{
\code{USYD_dIndex} is 446 row, 2 column \code{dataframe}. 
}
\details{This data frame is a typical soil information table where observations are matched prediction predictions from a specified model.}

\references{
\itemize{
\item Malone, B.P., Minasny, B., McBratney, A.B. (2017) \href{https://link.springer.com/book/10.1007/978-3-319-44327-0}{Using R for Digital Soil Mapping}. Springer Cham. 262 Pages.
\item Malone, B.P.,  McBratney, A.B., Minasny, B. (2018) \href{https://doi.org/10.7717/peerj.4659}{Description and spatial inference of soil drainage using matrix soil colours in the Lower Hunter Valley, New South Wales, Australia.}. PeerJ 6, e4659.

}
}

\examples{

# library(ithir)
# data(USYD_dIndex)
# summary(USYD_dIndex)

}
\keyword{datasets}
