\name{USYD soil data}
\docType{data}
\alias{USYD soil data}
\title{Random selection of soil point data}
\description{A soil information \code{dataframe} for 29 soil profiles. Each row is a horizon or depth interval observation for a given soil profile. Various soil attribute information is attibuted to each observation }
\usage{data(USYD_soil1)}
\format{
\code{USYD_soil1} is 166 row \code{dataframe} with identifier, soil depth and soil attribute information labelled in respective columns. 
}
\details{This data frame is a typical soil information table}

\references{
\itemize{

\item Malone, B.P., Minasny, B., McBratney, A.B. (2017) \href{https://link.springer.com/book/10.1007/978-3-319-44327-0}{Using R for Digital Soil Mapping}. Springer Cham. 262 Pages.
}
}

\examples{

# library(ithir)
# data(USYD_soil1)
# summary(USYD_soil1)

}
\keyword{datasets}
